worker/requirements.txt
redis
python-dotenv
ollama
